pyproject.toml
[project]
name = "course-planner-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.4.0",
    "bitsandbytes>=0.37.0",
    "chromadb>=0.6.3",
    "fastapi[standard]>=0.115.8",
    "httpx>=0.28.1",
    "langchain>=0.3.19",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.18",
    "langchain-huggingface>=0.1.2",
    "langchain-mongodb>=0.5.0",
    "langchain-openai>=0.3.7",
    "langgraph>=0.3.2",
    "langgraph-checkpoint-mongodb>=0.1.1",
    "numpy>=1.26.4",
    "pymongo>=4.11.1",
    "sentence-transformers>=3.4.1",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "iprogress>=0.4",
    "ipykernel>=6.29.5",
    "ipython>=8.32.0",
    "mypy>=1.15.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.7",
    "tqdm>=4.67.1",
]
